// src/app/common/constant/navigation-route.enum.ts
export enum NavigationRoute {
  ABOUT_US = 'about-us',
  ARTICLES_AND_ADVICE = 'articles-and-advice',
  BLOG = 'blog',
  CONTACT = 'contact',
  FOR_CONTRACTORS = 'for-contractors',
  FULL = 'full',
  HOMEPAGE = 'home',
  HOMEPAGE_REDIRECT = '/home',
  HOW_IT_WORKS = 'how-it-works',
  NEWS = 'news',
  PRIVACY_POLICY = 'privacy-policy',
  SERVICES = 'services',
  TERMS_OF_SERVICE = 'terms-of-service',
  WILDCARD = '*',
  EMPTY = '',
  FACEBOOK = 'https://www.facebook.com/fixmiyard',
  INSTAGRAM = 'https://www.instagram.com/fixmiyard'
}